+++ leaflet-plugins/leaflet-gesture-handling-1.4.2/leaflet-gesture-handling.js
@@ -146,7 +146,6 @@
 			(this._isLanguageContent(opts.text) ? Promise.resolve(opts.text) : this._getLanguageContent(opts.locale)).then((content) => {
 				this._map._container.setAttribute("data-gesture-handling-touch-content", content.touch);
 				this._map._container.setAttribute("data-gesture-handling-scroll-content", content.scroll);
-		
 				this._touchWarning = content.touch;
 				this._scrollWarning = content.scroll;
 			});
@@ -160,7 +159,6 @@
 		_getLanguageContent: function(lang) {
 			//Determine user language (eg. fr or en-US)
 			lang = lang || this._getUserLanguage() || "en";
-			
 			var resolve, promise = new Promise(_resolve => { resolve = _resolve; });
 			var consume = (m) => {
 				var content = m.default || {};
@@ -180,18 +178,9 @@
 			return promise;
 		},
 
-		_hasClass: function(element, classList) {
-			for (var i = 0; i < classList.length; i++) {
-				if (L.DomUtil.hasClass(element, classList[i])) {
-					return true;
-				}
-			}
-			return false;
-		},
-
 		_handleTouch: function(e) {
-			//Disregard touch events on the minimap if present
-			var ignore = this._hasClass(e.target, ["leaflet-control-minimap", "leaflet-interactive", "leaflet-popup-content", "leaflet-popup-content-wrapper", "leaflet-popup-close-button", "leaflet-control-zoom-in", "leaflet-control-zoom-out"]);
+			//Disregard touch events on the controls and popups if present
+			var ignore = L.DomUtil.hasClass(e.target, "leaflet-interactive") || e.target.closest('.leaflet-control-container') || e.target.closest('.leaflet-popup-pane');
 
 			if (ignore) {
 				if (L.DomUtil.hasClass(e.target, "leaflet-interactive") && e.type === "touchmove" && e.touches.length === 1) {