+++ changes.md
@@ -1,3 +1,3 @@
 * Solved Bug: detect Browser language in gesture handling
+* leaflet-gesture-handling-1.4.2
 * leaflet-elevation-2.2.6
-* leaflet-gesture-handling: leaflet-control-* (code opt.)